etl/python/requirements.txt
pandas>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
tqdm>=4.65.0
